integration complete

+++ src/headers/User.hpp
@@ -6,15 +6,15 @@ class User{
     private:
         string _name;
         int _accountNumber;  //for the sake of simplicity defining account number as interger.
-        int _accountBalance;    //for the sake of simplicity defining balance as integer
+        float _accountBalance;    //for the sake of simplicity defining balance as integer
     public:
         User();
         string const getName();
         void setName(const string name);
         int const getAccountNumber();
         void setAccountNumber(const int accountNumber);
-        int const getAccountBalance();
-        void setAccountBalance(const int accountBalance);
+        float const getAccountBalance();
+        void setAccountBalance(const float accountBalance);
         ~User();
 };
